working on task 15.

+++ src/comp1110/ass2/Card.java
@@ -18,7 +18,7 @@ public class Card extends ImageView {
         char[] chars = str.toCharArray();
 
         for (int i = 0; i < chars.length; i++) {
-            double x = ((double) i % 3) * SQUARE_WIDTH;
+            double x = (i % 3) * SQUARE_WIDTH;
             double y = (i / 3) * SQUARE_WIDTH;
             Square current = new Square(x, y, chars[i]);
